requirements.txt
pandas
xlrd # data is in excel format
numpy
scikit-learn
imbalanced-learn
omegaconf
pytest
tensorflow-cpu
mlflow
holidays